Changed the type of EventSequence from Long to Integer because Longs are annoying and there's no need for such a large number (since the sequence number is per unique Aggregate)

+++ src/main/java/dev/ted/jitterticket/eventsourced/domain/Event.java
@@ -1,5 +1,5 @@
 package dev.ted.jitterticket.eventsourced.domain;
 
 public interface Event {
-    Long eventSequence();
+    Integer eventSequence();
 }
