+++ src/main/java/dev/ted/jitterticket/eventsourced/domain/concert/ConcertScheduled.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 public record ConcertScheduled(ConcertId concertId,
-                               Long eventSequence,
+                               Integer eventSequence,
                                String artist,
                                int ticketPrice,
                                LocalDateTime showDateTime,
