+++ src/main/java/dev/ted/jitterticket/eventsourced/domain/customer/CustomerRegistered.java
@@ -1,7 +1,7 @@
 package dev.ted.jitterticket.eventsourced.domain.customer;
 
 public record CustomerRegistered(CustomerId customerId,
-                                 Long eventSequence,
+                                 Integer eventSequence,
                                  String customerName,
                                  String email)
         implements CustomerEvent {
